BL-11344: remove batch processing and introduce DLQ
Prior to this change the notifier lambda pulled messages from SQS in batches of
10 and processed them across 10 threads. The multithreading caused some problems
in production: if the details of an HGV could not be retrieved from TAPI that
message could not be processed; the multithreading meant that the failure mode
was non-deterministic; the batching meant that 10 messages would go back onto
the queue. Much sadness.

This change removes the batching and multithreading. Single messages are now
pushed to the notifier lambda via AWS event source mapping. There are
corresponding Terraform changes in webops/motr.